Commit para não perder nada.

+++ src/main/csharp/WebServer/Handlers/Controller/ViewResult.cs
@@ -1,4 +1,7 @@
 using System;
+using System.IO;
+using System.Reflection;
+using System.Text;
 using System.Web;
 
 namespace WebServer.Handlers.Controller
@@ -15,7 +18,7 @@ namespace WebServer.Handlers.Controller
         private readonly string _contentType;
         private readonly string _content;
 
-        public ViewResultBase(string content, string contentType)
+        protected ViewResultBase(string content, string contentType)
         {
             _content = content;
             _contentType = contentType;
@@ -41,21 +44,47 @@ namespace WebServer.Handlers.Controller
             DoWriteContent(response);
         }
 
-        protected abstract void DoWriteContent(HttpResponse response);
+        protected virtual void DoWriteContent(HttpResponse response)
+        {
+            response.Write(GetContent());
+        }
     }
 
     public class ViewResult : ViewResultBase
     {
+        private readonly bool _isFile = true;
+
         public ViewResult(string fileName) : base(fileName, "text/html")
         {
 
+        }
+
+        public ViewResult(string fileName, params string[] param)
+            : base(GenerateCompletePage(fileName, param), "text/html")
+        {
+            _isFile = false;
+        }
+
+        private static string GenerateCompletePage(string fileName, params string[] param)
+        {
+            StreamReader reader = new StreamReader(fileName);
+            string content = reader.ReadToEnd();
+
+            string.Format(content, param);
+
+            return content;
         }
 
         #region Overrides of ViewResultBase<IntPtr>
 
         protected override void DoWriteContent(HttpResponse response)
         {
-            response.WriteFile(GetContent());
+            if(_isFile)
+                response.WriteFile(GetContent());
+            else
+            {
+                response.Write(GetContent());
+            }
         }
 
         #endregion
@@ -67,14 +96,53 @@ namespace WebServer.Handlers.Controller
         {
             
         }
+    }
 
-        #region Overrides of ViewResultBase
-
-        protected override void DoWriteContent(HttpResponse response)
+    public class JsonResult<T> : ViewResultBase
+    {
+        public JsonResult(T obj) : base(ParseToJson(obj), "application/json")
         {
-            response.Write(GetContent());
+
         }
 
-        #endregion
+        public JsonResult(T[] arrayObj)
+            : base(ParseArrayToJson(arrayObj), "application/json")
+        {
+            
+        }
+
+        private static string ParseArrayToJson(T[] arrayObj)
+        {
+            StringBuilder json = new StringBuilder("[");
+            
+            foreach(T obj in arrayObj)
+            {
+                json.AppendFormat("{0}, ", ParseToJson(obj));
+            }
+            json.Remove(json.Length - 1, 1);
+            json.Append("]");
+
+            return json.ToString();
+        }
+
+        private static string ParseToJson(T o)
+        {
+            Type typeObj = o.GetType();
+            StringBuilder json = new StringBuilder("{");
+
+            foreach (var property in typeObj.GetProperties(BindingFlags.Public | BindingFlags.Instance))
+            {
+                json.AppendFormat("{0} : {1},", property.Name, property.GetValue(o, null));
+            }
+
+            foreach (var field in typeObj.GetFields(BindingFlags.Public | BindingFlags.Instance))
+            {
+                json.AppendFormat("{0} : {1},", field.Name, field.GetValue(o));
+            }
+
+            json.Remove(json.Length - 1, 1);
+            json.Append("}");
+            return json.ToString();
+        }
     }
 }
